Update April 6
I removed figures, changed footnotes to appear how they do in B1. These might need tweaking to appear just as they do in the facsimile, but I wanted to be sure it showed up in HTML for proofing. Note that sentences should not be broken by <p> because <p> should be used exclusively for when the facsimile has a paragraph break. Use B1 or other narratives as a coding reference for that.