Update April 10
I hopefully fixed the hierarchy issues. To clarify for everyone, the hierarchy issues are really just issues with the <head> and <div> divisions at the start of the translated part. There are also footnote issues which we may need to return to. I've been changing them to how they appear in B1, because I want them to definitely appear in the HTML proof. From there, we can change them if we want them to be exactly like the periodical appears. (I think the periodical uses superscript, where right now the footnote number is italicized).